Do not hang in timer synchronization init without optimizations.
There is an operation that must be performed within 1 RTC cycle to make
sure that synchronization would be performed in time. There was an
infinite loop that iterated until operation was performed within 1 RTC
cycle. The problem was that without compiler optimizations single
iteration takes more than 1 RTC cycle. It caused hanging of
initialization procedure.

This patch provides a counter that limits number of iterations of this
loop. If the limit is reached, the operation is not performed again and
the timers are most probably unsynchronized. It is assumed that having
unsynchronized timers is better than hanging up the driver during debug
sessions.